+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Apply policies to v20 even when v20 contains suffixes in its name.
 
+### Added
+
+- Policy to apply `audit-policy.yaml` to kubeadmconfig
+
 ## [0.9.2] - 2021-10-26
 
 ### Changed